+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## next ##
 
 * Viewer: Show release details when prompting to update
+* Viewer: Fix invalid thread error when new update detected at app start
 
 ## v0.12.4 ##